+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.0] — 2023-09-20
+
+- Feature: limit homepage and sidebar to current course terms
+- Feature: add page template to display all courses for multiple terms
+- Feature: add page template to display all user favorites
+- Feature: add logout link to main navigation menu
+
 ## [1.5.1] — 2023-08-03
 
 - Bugfix: fix guides display